feat: complete JavaScript to TypeScript migration
- Remove all legacy JavaScript files
- Update Dockerfile for TypeScript build process
- Clean up obsolete service files and configurations

Removed Files:
- Legacy JavaScript server and middleware
- Old proxy implementations
- Outdated utility functions
- Deprecated service configurations

Docker Changes:
- Multi-stage build process for TypeScript compilation
- Optimized image size with production dependencies only
- Updated build commands for TypeScript workflow

Migration Complete:
✅ Full TypeScript conversion with type safety
✅ Enhanced error handling and logging
✅ Improved security and middleware stack
✅ Comprehensive documentation
✅ Production-ready Docker configuration
✅ Enterprise-grade development workflow
