+++ api-gateway/Dockerfile
@@ -8,6 +8,6 @@ RUN npm ci --only=production
 
 COPY . .
 
-EXPOSE 3000
+EXPOSE 4000
 
 CMD ["node", "src/server.js"]
