--- identity-service/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "identity-service",
-  "version": "1.0.0",
-  "main": "src/server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/server.js",
-    "dev": "nodemon src/server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "argon2": "^0.41.1",
-    "cors": "^2.8.5",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
-    "express-rate-limit": "^7.5.0",
-    "helmet": "^8.1.0",
-    "ioredis": "^5.6.1",
-    "joi": "^17.13.3",
-    "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.13.2",
-    "rate-limit-redis": "^4.2.0",
-    "rate-limiter-flexible": "^7.1.0",
-    "winston": "^3.17.0"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.9"
-  }
-}
