--- identity-service/src/utils/redis.js
@@ -1,5 +1,0 @@
-const Redis = require("ioredis");
-
-const clientRedis = new Redis(process.env.REDIS_URL);
-
-module.exports = clientRedis;
