--- post-service/Dockerfile
@@ -1,13 +1,0 @@
-FROM node:18-alpine
-
-WORKDIR /usr/src/app
-
-COPY package*.json ./
-
-RUN npm ci --only=production
-
-COPY . .
-
-EXPOSE 3002
-
-CMD ["node", "src/server.js"]
