--- post-service/src/config/constants.js
@@ -1,7 +1,0 @@
-module.exports = {
-  CACHE_TTL: {
-    POSTS_LIST: 300,
-    SINGLE_POST: 3600,
-  },
-  DEFAULT_PAGE_LIMIT: 10,
-};
