--- post-service/src/middleware/authMiddleware.js
@@ -1,19 +1,0 @@
-const logger = require("../utils/logger");
-
-const authenticateRequest = (req, res, next) => {
-  const userId = req.headers["x-user-id"];
-
-  if (!userId) {
-    logger.warn("Access attempted without user ID");
-    return res.status(401).json({
-      success: false,
-      message: "Authencation required! Please login to continue",
-    });
-  }
-
-  req.user = { userId };
-
-  next();
-};
-
-module.exports = authenticateRequest;
