--- post-service/src/models/Post.js
@@ -1,33 +1,0 @@
-const mongoose = require("mongoose");
-
-const PostSchema = new mongoose.Schema(
-  {
-    user: {
-      type: mongoose.Schema.Types.ObjectId,
-      ref: "User",
-      required: true,
-    },
-    content: {
-      type: String,
-      required: true,
-    },
-    mediaIds: [
-      {
-        type: String,
-      },
-    ],
-    createdAt: {
-      type: Date,
-      default: Date.now,
-    },
-  },
-  {
-    timestamps: true,
-  }
-);
-
-PostSchema.index({ content: "text" });
-
-const Post = mongoose.model("Post", PostSchema);
-
-module.exports = Post;
